requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
whitenoise==6.8.2
python-decouple==3.8
requests==2.32.3
